Fix: #592 ディレクトリを変更してAppDialogを適用させる
